add items.

+++ Zbang.Cloudents.Mvc4WebRole/Js/components/quiz/quizCreate.close.controller.js
@@ -18,9 +18,13 @@
         }
 
         function del() {
-            quizService.deleteQuiz(quizId).then(function() {
-                draft();
-            });
+            if (quizId) {
+                quizService.deleteQuiz(quizId).then(function() {
+                    draft();
+                });
+                return;
+            }
+            draft();
         }
     }
 })()
